requirements.txt
python-chess
flask-restplus